[Refactor] Convert NodeBadge.vue to composable
